requirements-dev.txt
# Incluir dependências principais
-r requirements.txt

# Testes
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Empacotamento
PyInstaller==6.0.0

# Linting e Formatação
flake8==6.1.0
black==23.9.1
isort==5.12.0

# Debug
Flask-DebugToolbar==0.13.1